Cont building update_owned_list route
If game is already on list or the game is in the steam_games table then
the correct behavior currently exists.  If neither of these is true,
then current behavior is to get the app data, but it does not yet add
the game to the user's owned list - need to add this behavior next.